添加排除词语;工具类同意使用;优化词云

+++ src/main/java/com/haruhi/bot/dispenser/MessageDispenser.java
@@ -217,7 +217,7 @@ public class MessageDispenser {
         IMessageEventType messageEventType;
         try {
             funId = Integer.valueOf(fun);
-            messageEventType = findHandlerByWeight(funId.intValue());
+            messageEventType = findHandlerByWeight(funId);
         }catch (Exception e){
             messageEventType = findHandlerByName(fun);
         }
